Se cambio las variables de vida de int a float
se cambio de tipo de variable para hacer mas preciso el calculo de nivel

+++ PROYECTO EMILY REP/Assets/Scripts/Character/CharacterUIManager.cs
@@ -8,7 +8,7 @@ namespace KC
         public bool hasFloatingHPBar= true;
         public UI_Character_HP_Bar characterHpBar;
 
-        public void OnHPChanged(int oldValue, int newValue)
+        public void OnHPChanged(float oldValue, float newValue)
         {
             characterHpBar.oldHealthValue = oldValue;
             characterHpBar.SetStat(newValue);
